Day 17 Solution and updates for Point to long (#20)

+++ src/main/java/net/ddellspe/day15/Day15.java
@@ -30,10 +30,10 @@ public class Day15 {
     for (Entry<Point, Point> entry : sensors.entrySet()) {
       Point sensor = entry.getKey();
       Point beacon = entry.getValue();
-      int distance = sensor.getManhattanDistance(beacon);
+      long distance = sensor.getManhattanDistance(beacon);
       if ((sensor.getY() - distance) <= yLevel && (sensor.getY() + distance) >= yLevel) {
-        int yDiff = distance - Math.abs(yLevel - sensor.getY());
-        for (int x = -yDiff; x <= yDiff; x++) {
+        long yDiff = distance - Math.abs(yLevel - sensor.getY());
+        for (long x = -yDiff; x <= yDiff; x++) {
           nonBeacons.add(new Point(sensor.getX() + x, yLevel));
         }
       }
@@ -45,28 +45,28 @@ public class Day15 {
 
   public static long part2(String filename, int maxCoord) {
     List<String> lines = InputUtils.stringPerLine(filename, Day15.class);
-    Map<Point, Integer> sensors = new LinkedHashMap<>();
+    Map<Point, Long> sensors = new LinkedHashMap<>();
     for (String line : lines) {
-      int sensorX = Integer.parseInt(line.substring(line.indexOf("x=") + 2, line.indexOf(", ")));
-      int sensorY = Integer.parseInt(line.substring(line.indexOf("y=") + 2, line.indexOf(":")));
-      int beaconX =
+      long sensorX = Integer.parseInt(line.substring(line.indexOf("x=") + 2, line.indexOf(", ")));
+      long sensorY = Integer.parseInt(line.substring(line.indexOf("y=") + 2, line.indexOf(":")));
+      long beaconX =
           Integer.parseInt(
               line.substring(
                   line.indexOf("x=", line.indexOf(":")) + 2,
                   line.indexOf(", ", line.indexOf(":"))));
-      int beaconY = Integer.parseInt(line.substring(line.indexOf("y=", line.indexOf(":")) + 2));
+      long beaconY = Integer.parseInt(line.substring(line.indexOf("y=", line.indexOf(":")) + 2));
       Point sensor = new Point(sensorX, sensorY);
       sensors.put(sensor, sensor.getManhattanDistance(new Point(beaconX, beaconY)));
     }
     Point testPoint = null;
-    for (Entry<Point, Integer> pointDistances : sensors.entrySet()) {
-      for (int yOffset = -pointDistances.getValue() - 1;
+    for (Entry<Point, Long> pointDistances : sensors.entrySet()) {
+      for (long yOffset = -pointDistances.getValue() - 1;
           yOffset < pointDistances.getValue() + 1;
           yOffset++) {
-        int xOffset = pointDistances.getValue() - Math.abs(yOffset) + 1;
-        int minX = pointDistances.getKey().getX() - xOffset;
-        int maxX = pointDistances.getKey().getX() + xOffset;
-        int y = pointDistances.getKey().getY() + yOffset;
+        long xOffset = pointDistances.getValue() - Math.abs(yOffset) + 1;
+        long minX = pointDistances.getKey().getX() - xOffset;
+        long maxX = pointDistances.getKey().getX() + xOffset;
+        long y = pointDistances.getKey().getY() + yOffset;
         if (minX >= 0
             && minX <= maxCoord
             && y >= 0
